job_platform_app/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-graphql-jwt==0.4.0
django-location-field==2.7.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
inflection==0.5.1
packaging==25.0
pillow==12.0.0
promise==2.3
psycopg==3.2.12
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
uritemplate==4.2.0
